pyproject.toml
[project]
name = "gerby-web"
version = "0.0.1"
description = "."
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "xmldiff>=2.6.3",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
gerbyTest = "gerbyTester.cli:cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
